database_creation/requirements.txt
sentencepiece
blobfile
protobuf
bitsandbytes
einops
accelerate
setuptools-rust
openai-whisper
pytubefix
pydub
pandas